[fix] upgrade version -> 0.1.0

+++ iToys-kit/core/src/main/kotlin/com/itoys/android/core/CoreConst.kt
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
  */
 data class DictModel(
     @JsonProperty("dictKey")
-    val dictKey: Long?,
+    val dictKey: Int?,
     @JsonProperty("dictValue")
     val dictValue: String?,
 ) {
